Update event checkbox handling
This is to support storing the correct set of FRMs for creating movies.
Not all FRMs may appear on screen.
When selecting all of them, we should only include what's on screen.
The previous implementation always used "all" frms when all boxes were checked.
This results in the backend using "all" frms even though not all frms were really selected.